Server: correctly handle duplicate AppendEntriesResults
Because this implementation uses messages and not true RPC, it doesn't
inherently match up messages with responses, so when a server
retransmits a request, the recipient may send back multiple responses.

In order to prevent multiple responses to the same AppendEntries from
being mistaken to be responses to separate AppendEntries, include the
match index in the results and remove internal tracking of the number
of log entries actually sent in the last AppendEntries.