Server: cleanup, possible performance improvement
Drop state of any server no longer in the cluster.  Otherwise, it may
be possible the the collection will grow in size as we add and remove
more and more servers, to the point where we're keeping a lot of state
we don't need, and are having to constantly enumerate over it.