Fix bug with match index in AppendEntriesResults
The bug was that if a follower received an AppendEntries with
entries all before the last index the follower has, the follower
would return a match index beyond the entries sent by the leader.

The solution is to return the follower's last index or the index
of the last entry covered by the AppendEntries, whichever is less.