Fix bug involving responses to retransmitted requests
The bug was that if a request was retransmitted and the other
server responded to both, the leader would act on both responses,
causing problems such as installing a snapshot or attempting log
replication while another request was in progress.

The solution is to use message sequence numbers, and ignore responses
whose sequence numbers, if present, do not match the sequence numbers
of the last requests sent.