Fix race condition handling timeouts
A fixed number of timeouts are always registered
and then cancelled since each state change causes
the task scheduling timeouts to cancel them.

We only want to actually complete the last timeout
registered, but we have hold onto the others so that
they aren't prematurely completed.  So collect them
all and complete the last one first.  After that we
can just drop the rest; it doesn't matter if they
complete or not, because they should be cancelled by
the time the last is registered.