Fix problem with mock timers whose senders get dropped
If a mock timer sender is dropped, the await on the receiver
would complete (with error), which would cause the Raft
code to think that the timers were triggered/expired.

Fix this by detecting when the sender is dropped (in this case
the `receiver.await` returns an error), and when this happens,
simply never complete the future at all.