Redesign how log compaction is communicated
The Raft server now calls a new method `Log::update_snapshot`
when the host should update the snapshot.  The host should
not change the snapshot, base index, or base term at any point
other than during the call to `update_snapshot`.  It's expected
that the majority of the log compaction work is done in a
separate thread or task, and merely moved or copied when
`update_snapshot` is called.